Change list and get request from s3 to db

+++ controllers/commands.py
@@ -8,7 +8,7 @@ import pytz
 from settings import TELEGRAM_ADMIN_USERNAME
 
 from services.aws import list_folders, get_hq, list_hqs
-from services.db import insert_publishers, insert_titles, update_editions
+from services.db import insert_publishers, insert_titles, update_editions, select
 from services.telegram import alert_admin
 
 def start(update: Update, context: CallbackContext):
@@ -30,12 +30,13 @@ def help(update: Update, context: CallbackContext):
 
 
 def list_publishers(update: Update, context: CallbackContext):
-    publishers = list_folders('')
+    publishers = select('publishers')
+    print(publishers)
     keyboard = []
     print('-------------------')
     print('Editoras:')
     for publisher in publishers:
-        publisher_name = str(publisher).replace('/', '')
+        publisher_name = publisher[1]
         print(publisher_name.upper())
         keyboard.append([InlineKeyboardButton(publisher_name.upper(), callback_data=str(publisher_name))])
     print('-------------------')
@@ -50,28 +51,30 @@ def list_editions(update: Update, context: CallbackContext):
     timezone = pytz.timezone("America/Sao_Paulo")
 
     if len(context.args) == 0:
-        update.message.reply_text("""Para descobrir a quantidade de edições de uma HQ, é necessário passar a editora e o título.
-        \n Exemplo: /edicoes MARVEL A_THOR""")
+        update.message.reply_text("""Para descobrir a quantidade de edições de uma HQ, é necessário passar o título.
+        \n Exemplo: /edicoes A_THOR""")
         return
-    if len(context.args) != 2:
+    if len(context.args) != 1:
         update.message.reply_text("""⚠️ Opa, calma aí! ⚠️
         \n Verifique se mandou tudo direitinho.
-        \n Você precisa passar 2 informações: a editora e o título.
+        \n Você precisa passar apenas uma informação: o título.
         \n Não esqueça que o nome do título deve ter _ (underline) no lugar dos espaços :b""")
         return
-    editions = list_hqs(f'{context.args[0]}/{context.args[1]}')
+
+    editions = select('editions', 'title', context.args[0])
+    print(editions)
     if not type(editions).__name__ == 'list':
         print(f'Não achei a lista de edições. Retornou: {editions}')
-        update.message.reply_text(f'{editions}')
-        alert_admin(f'{name} (@{username}) Procurou por {context.args[0]}/{context.args[1]} às {datetime.datetime.now().astimezone(timezone).strftime("%d/%m/%Y %H:%M")} mas infelizmente não temos :(')
+        update.message.reply_text(f'Infelizmente não temos nenhuma edição desse título no momento :(')
+        alert_admin(f'{name} (@{username}) Procurou por {context.args[0]} às {datetime.datetime.now().astimezone(timezone).strftime("%d/%m/%Y %H:%M")} mas infelizmente não temos :(')
         return
     title = context.args[1].upper().replace('_', ' ')
     message = f'Nós temos {len(editions)} edições de {title}: \n'
     print('-------------------')
     print('Edições:')
     for edition in editions:
-        message = message + f'\n {title} #{edition}.pdf'
-        print(f'{title} #{edition}.pdf')
+        message = message + f'\n {title} #{edition[1]}.pdf'
+        print(f'{title} #{edition[1]}.pdf')
     update.message.reply_text(f'{message}')
     print('-------------------')
 
@@ -82,49 +85,53 @@ def download(update: Update, context: CallbackContext):
     timezone = pytz.timezone("America/Sao_Paulo")
 
     if len(context.args) == 0:
-        update.message.reply_text("""Para baixar uma HQ é necessário passar a editora, o título e a edição 
-        \n Exemplo: /baixar MARVEL A_THOR 1""")
+        update.message.reply_text("""Para baixar uma HQ é necessário passar o título e a edição 
+        \n Exemplo: /baixar A_THOR 1""")
         return
-    if len(context.args) != 3:
+    if len(context.args) != 2:
         update.message.reply_text("""⚠️ Opa, calma aí! ⚠️
         \n Verifique se mandou tudo direitinho.
-        \n Você precisa passar 3 informações: a editora, o título e o número da edição.
+        \n Você precisa passar 2 informações: o título e o número da edição.
         \n Não esqueça que o nome do título deve ter _ (underline) no lugar dos espaços :b""")
         return
-    publisher = (context.args[0]).lower()
-    title = (context.args[1]).lower().replace('-', '_')
-    edition = context.args[2]
-    print(f'Parâmentros enviados para download: {publisher}, {title}, {edition}')
-    try: 
-        get_hq(publisher, title, edition)
-    except Exception as error:
-        print(error)
-        if "404" in str(error):
-            update.message.reply_text('Poxa, ainda não temos esse título disponível para download :(')    
-        else:
+
+    title = (context.args[0]).lower().replace('-', '_')
+    edition = context.args[1]
+    print(f'Parâmentros enviados para download: {title}, {edition}')
+    print(select('editions', ['identifier', 'title'], [edition, title]))
+    if select('editions', ['identifier', 'title'], [edition, title]):
+        publisher = select('hqs', 'name', title)
+        try: 
+            get_hq(publisher[0][2], title, edition)
+        except Exception as error:
+            print(error)
             update.message.reply_text('Sinto muito! Tivemos um erro ao buscar esse título :(')
-        alert_admin(f'{name} (@{username}) Procurou por {publisher}/{title}/{edition} às {datetime.datetime.now().astimezone(timezone).strftime("%d/%m/%Y %H:%M")} mas infelizmente não temos :(')
-        return error 
-    if os.path.exists(f'data/{title}_{edition}.pdf'):
-        update.message.reply_text('Opa! Achei aqui. Só um segundo que já estou enviando...')
-        hq_path = f'data/{title}_{edition}.pdf'
-        new_title = f'{title.replace("_", " ").title()} #{edition}.pdf'
-        update.message.reply_document(document=open(hq_path, 'rb'), filename=new_title, timeout=1000)
-    update.message.reply_text('Prontinho. Divirta-se! :D')
-    print(f'HQ -> {new_title} Enviada!')
-    print('-------------------')
-    os.remove(hq_path)
-    alert_admin(f'{name} (@{username}) Solicitou {new_title} às {datetime.datetime.now().astimezone(timezone).strftime("%d/%m/%Y %H:%M")}')
+            alert_admin(f'{name} (@{username}) Procurou por {publisher[0][2]}/{title}/{edition} às {datetime.datetime.now().astimezone(timezone).strftime("%d/%m/%Y %H:%M")} mas infelizmente não conseguimos enviar :(')
+            return error 
+        if os.path.exists(f'data/{title}_{edition}.pdf'):
+            update.message.reply_text('Opa! Achei aqui. Só um segundo que já estou enviando...')
+            hq_path = f'data/{title}_{edition}.pdf'
+            new_title = f'{title.replace("_", " ").title()} #{edition}.pdf'
+            #update.message.reply_document(document=open(hq_path, 'rb'), filename=new_title, timeout=1000)
+        update.message.reply_text('Prontinho. Divirta-se! :D')
+        print(f'HQ -> {new_title} Enviada!')
+        print('-------------------')
+        os.remove(hq_path)
+        alert_admin(f'{name} (@{username}) Solicitou {new_title} às {datetime.datetime.now().astimezone(timezone).strftime("%d/%m/%Y %H:%M")}')
+    else:
+        update.message.reply_text('Poxa, ainda não temos esse título ou edição disponível para download :(')
+        alert_admin(f'{name} (@{username}) Procurou por {publisher[0][2]}/{title}/{edition} às {datetime.datetime.now().astimezone(timezone).strftime("%d/%m/%Y %H:%M")} mas infelizmente não temos :(')
 
 
 def sync(update: Update, context: CallbackContext):
-    alert_admin(f'Iniciando sincronização com o banco...')
-    update.message.reply_text('Iniciando sincronização com o banco...')
-
     username = update.message.from_user.username
+    name = update.message.from_user.full_name    
     timezone = pytz.timezone("America/Sao_Paulo")
 
     if username == TELEGRAM_ADMIN_USERNAME:
+        alert_admin(f'Iniciando sincronização com o banco...')
+        update.message.reply_text('Iniciando sincronização com o banco...')
+
         folders = list_folders('')
         publishers = []
         for publisher in folders:
@@ -158,7 +165,7 @@ def sync(update: Update, context: CallbackContext):
 
 
     else:
-        alert_admin(f'@{username}) tentou usar o comando de sincronização às {datetime.datetime.now().astimezone(timezone).strftime("%d/%m/%Y %H:%M")}')
+        alert_admin(f'{name} @({username}) tentou usar o comando de sincronização às {datetime.datetime.now().astimezone(timezone).strftime("%d/%m/%Y %H:%M")}')
         update.message.reply_text('Você não tem permissão para usar esse comando.')
